Create graph module, fix roads module
