+++ mailRoute.js
@@ -1,3 +1,5 @@
+// The best route, used for memory
 const MAILROUTE = ["Alice's House", "Cabin", "Alice's House", "Bob's House", "Town Hall", "Daria's House", "Ernie's House", "Grete's House", "Shop", "Grete's House", "Farm", "Marketplace", "Post Office"];
 
+// exporting to main file
 module.exports = MAILROUTE;
