+++ robot.js
@@ -1,37 +1,8 @@
-const ROADS = require("./roads");
+// Importing road graph
+const roadGraph = require("./roads");
+
+// memory fot the best routes
 const MAILROUTE = require("./mailRoute");
-// const buildGraph = require("./roads");
-// const efficientRobot = require("./efficient_robot");
-// console.log(MAILROUTE);
-
-//Building a graph, for each place tells what can be reached from there
-// edges - roads
-function buildGraph(edges) {
-  //creat an empty object 
-    let graph = Object.create(null);
-    function addEdge(from, to) {
-      if (graph[from] == null) {
-        graph[from] = [to];
-      } else {
-        graph[from].push(to);
-      }
-    }
-  // Iterating through array elements (map), and spliting it into 2 el.(split)
-    for (let [from, to] of edges.map(r => r.split("-"))) {
-  
-  // Adding places you can reached out to for the 1st place in the list
-      addEdge(from, to);
-  
-  // For the ssecond place destinations place will be the 1st place
-      addEdge(to, from);
-    }
-  // Returns an object with places names which contains an array of possible destinations
-    return graph;
-  }
-
-const roadGraph = buildGraph(ROADS);
-// module.exports = roadGraph;
-// console.log(roadGraph);
 
 //The Task
 class VillageState {
@@ -193,7 +164,6 @@ function compareRobots(robot1, memory1, robot2, memory2) {
   console.log(`Robot ${robot1.name} needed ${total1 / 100} steps per task`)
   console.log(`Robot ${robot2.name} needed ${total2 / 100}`)
 }
-
 // count steps
 function countSteps(state, robot, memory) {
   for (let steps = 0;; steps++) {
